A bit more testing.
